+++ packages/sui-ssr/server/template/index.js
@@ -1,27 +1,18 @@
-import path from 'path'
-import fs from 'fs'
-
-const APP_FILE = 'index.html'
 const APP_PLACEHOLDER = '<!-- APP -->'
 const HEAD_CLOSING_TAG = '</head>'
 const BODY_CLOSING_TAG = '</body>'
-let _memoizedAppParts = null
+
 /**
- * Loads public/index.html once (memoized) and returns an array with two items,
+ * Loads public/index.html once and returns an array with two items,
  * one for the head and one for the body. Html opening and closing tag will be present respectively on each part.
- * @return {array}
+ * @return {Array}
  */
-export const getTplParts = () => {
-  if (_memoizedAppParts) {
-    return _memoizedAppParts
-  }
-  const filePath = path.join(process.cwd(), 'public', APP_FILE)
-  const htmlTemplate = fs.readFileSync(filePath, 'utf8')
+export const getTplParts = req => {
+  let appParts = req.htmlTemplate.split(HEAD_CLOSING_TAG)
 
-  _memoizedAppParts = htmlTemplate.split(HEAD_CLOSING_TAG)
-  _memoizedAppParts[0] = `${_memoizedAppParts[0]}${HEAD_CLOSING_TAG}`
+  appParts[0] = `${appParts[0]}${HEAD_CLOSING_TAG}`
 
-  return _memoizedAppParts
+  return appParts
 }
 
 export class HtmlBuilder {}