+++ packages/sui-bundler/CHANGELOG.md
@@ -2,6 +2,37 @@
 
 All notable changes to this project will be documented in this file.
 
+<a name="5.15.0"></a>
+# 5.15.0 (2019-06-06)
+
+
+### Features
+
+* support react-router-dom link-package ([f534f37](https://github.com/SUI-Components/sui/commit/f534f37))
+
+
+
+<a name="5.14.0"></a>
+# 5.14.0 (2019-06-06)
+
+
+### Features
+
+* add onlyHash config ([ebad99b](https://github.com/SUI-Components/sui/commit/ebad99b))
+
+
+
+<a name="5.13.0"></a>
+# 5.13.0 (2019-05-27)
+
+
+### Bug Fixes
+
+* fix incompatibility issue when use link-loader ([c2d76f7](https://github.com/SUI-Components/sui/commit/c2d76f7))
+* upgrade to not unsecure dependency ([bafcaff](https://github.com/SUI-Components/sui/commit/bafcaff))
+
+
+
 <a name="5.12.0"></a>
 # 5.12.0 (2019-05-24)
 
@@ -874,6 +905,16 @@ with the old one
 ### BREAKING CHANGES
 
 * CLI has changed name
+
+
+
+<a name="3.4.0"></a>
+# 3.4.0 (2017-06-27)
+
+
+### Features
+
+* migrate from sui-studio-webpack ([1b58081](https://github.com/SUI-Components/sui/commit/1b58081))